refactor(KERN): remove legacy design fieldset
