+++ stories/form/Fieldset.stories.tsx
@@ -1,6 +1,6 @@
 import { reactRouterFormContext } from "~/../.storybook/reactRouterFormContext";
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { KernFieldset } from "~/components/kern/formElements/KernFieldset";
+import { KernFieldset } from "~/components/formElements/KernFieldset";
 import demoImage from "../assets/Rechtliche_Unterstuetzung.svg?no-inline";
 
 const meta = {
